Redefine architecture
