--- api/not_found.rb
@@ -1,8 +1,0 @@
-module Embryo
-  # Handles 404 errors
-  class NotFound < Grape::API
-    route :any, '*path' do
-      error! 'Not found.', 404
-    end
-  end
-end
