+++ db/seeds.rb
@@ -1,5 +1,5 @@
 names = %w(Male Female)
 
 names.each do |name|
-  Gender.create!(name: name)
+  Embryo::Gender.create!(name: name)
 end
