--- models/gender.rb
@@ -1,6 +1,0 @@
-# Stores a person's gender
-class Gender < ActiveRecord::Base
-  validates :name, presence: true
-
-  has_many :person
-end
