+++ spec/embryo/models/embryo/gender_spec.rb
@@ -1,6 +1,6 @@
 require 'spec_helper'
 
-describe Gender, type: :model do
+describe Embryo::Gender, type: :model do
   it { is_expected.to validate_presence_of(:name) }
   it { is_expected.to have_many(:person) }
 end
