+++ spec/embryo/models/embryo/person_spec.rb
@@ -1,6 +1,6 @@
 require 'spec_helper'
 
-describe Person, type: :model do
+describe Embryo::Person, type: :model do
   it { is_expected.to validate_presence_of(:name) }
   it { is_expected.to validate_presence_of(:born_at) }
   it { is_expected.to belong_to(:gender) }
