+++ spec/support/factories/person_factory.rb
@@ -1,5 +1,5 @@
 FactoryGirl.define do
-  factory :person do
+  factory :person, class: 'Embryo::Person' do
     name 'Wilson Silva'
     born_at '15/11/1990'
   end
