+++ spec/web/embryo/people_api_spec.rb
@@ -1,13 +1,13 @@
 require 'spec_helper'
 
-describe Embryo::People, type: :api do
+describe Embryo::PeopleAPI, type: :api do
   describe 'POST /people' do
     it 'creates a person' do
       post '/people', {name: 'Big Brother', born_at: '04/04/1984'}, 'HTTP_ACCEPT' => 'application/vnd.embryo-v1+json'
 
       expect_status(201)
       expect_json(name: 'Big Brother', born_at: '1984-04-04T00:00:00.000+00:00')
-      expect(Person.first).to have_attributes(name: 'Big Brother')
+      expect(Embryo::Person.first).to have_attributes(name: 'Big Brother')
     end
   end
 
@@ -42,7 +42,7 @@ describe Embryo::People, type: :api do
 
       expect_status(204)
       expect(response.body).to be_empty
-      expect(Person.where(id: person.id)).not_to exist
+      expect(Embryo::Person.where(id: person.id)).not_to exist
     end
   end
